+++ src/game/verbs/keywords.ts
@@ -1,13 +1,13 @@
 import { Verb, newVerb } from "./verb";
 import { selectInventory, selectKeywords } from "../../utils/selectors";
 import { RandomText } from "../interactions/text";
-import { OptionGraph } from "../interactions/optionGraph";
 import { getHelp, giveHint } from "../../gonorth";
 import { getKeywordsTable } from "../../utils/defaultHelp";
 import { getStore } from "../../redux/storeRegistry";
 import { addKeywords, removeKeywords } from "../../redux/gameActions";
 import { handleDebugOperations } from "../../utils/debugFunctions";
 import { clearPage } from "../../utils/sharedFunctions";
+import createWaitGraph from "./waitGraph";
 
 export function createKeywords() {
   const emptyInventoryText = new RandomText(
@@ -35,38 +35,7 @@ export function createKeywords() {
     "Inspect the items you're carrying."
   );
 
-  const waitText = new RandomText(
-    "You pause for a moment, taking stock of the situation.",
-    "You look around you and drink in your surroundings, letting the moment linger.",
-    "You stop and think. Yes, you're sure there's a way out of this mess.",
-    "You wonder aloud whether you'll make it home alive. No-one answers.",
-    "How did you get yourself into this pickle? And how to clamber out?"
-  );
-
-  const stopWaitingText = new RandomText(
-    "You snap out of your reverie and return your attention to the room.",
-    "You've waited long enough. You get back to the task at hand.",
-    "Satisfied everything's in order, you consider what to do next."
-  );
-
-  const waitNodes = [
-    {
-      id: "wait",
-      actions: waitText,
-      options: {
-        "Keep waiting": "wait",
-        "Stop waiting": "stop"
-      }
-    },
-    {
-      id: "stop",
-      noEndTurn: true,
-      actions: stopWaitingText
-    }
-  ];
-
-  const waitGraph = new OptionGraph("wait", ...waitNodes);
-  const wait = new Verb("wait", true, waitGraph.commence(), [], [], true, "Allow time to pass.");
+  const wait = new Verb("wait", true, createWaitGraph().commence(), [], [], true, "Allow time to pass.");
 
   const help = new Verb(
     "help",